Cambio double monto pago Fix  Dialog AddPay Fix Validacion precio AddProduct

+++ app/src/main/java/com/example/inventariosapp/database/entity/PayEntity.kt
@@ -9,7 +9,7 @@ data class PayEntity(
     @PrimaryKey
     val ventaPagoId: Int,
     val ventaId: Int?,
-    val montoPago: Int?,
+    val montoPago: Double?,
     val fecha: String?,
     val observaciones: String?,
     val origenId: Int?,
